# scripts/openapi.yaml
openapi: '3.1.0'
info:
  title: Calendar Web App
  description: Web App to Display Calendar, Sync with GCal
  contact:
    name: Daniel Boline
    email: ddboline@gmail.com
  license:
    name: ''
  version: '0.8.4'
paths:
  /calendar/agenda:
    get:
      operationId: agenda
      responses:
        '200':
          description: Agenda
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /calendar/calendar_cache:
    get:
      operationId: calendar_cache
      responses:
        '200':
          description: Calendar Cache
          content:
            text/html:
              schema:
                type: object
                required:
                - pagination
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CalendarCache'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
    post:
      operationId: calendar_cache_update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarCacheUpdateRequest'
        required: true
      responses:
        '201':
          description: Calendar Cache Update
          content:
            text/html:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalendarCache'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /calendar/calendar_list:
    get:
      operationId: calendar_list
      responses:
        '200':
          description: Calendar List
          content:
            text/html:
              schema:
                type: object
                required:
                - pagination
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/_CalendarListWrapper'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
    post:
      operationId: calendar_list_update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarUpdateRequest'
        required: true
      responses:
        '201':
          description: Calendar List Update
          content:
            text/html:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/_CalendarListWrapper'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /calendar/create_calendar_event:
    get:
      operationId: build_calendar_event
      responses:
        '200':
          description: Build Calendar Event
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
    post:
      operationId: create_calendar_event
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_CreateCalendarEventRequest'
        required: true
      responses:
        '201':
          description: Create Calendar Event
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
  /calendar/delete_event:
    delete:
      operationId: delete_event
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GcalEventID'
        required: true
      responses:
        '204':
          description: Delete Event Output
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
  /calendar/event_detail:
    get:
      operationId: event_detail
      responses:
        '201':
          description: Event Details
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
components:
  schemas:
    CalendarCache:
      type: object
      required:
      - gcal_id
      - event_id
      - event_start_time
      - event_end_time
      - event_name
      - last_modified
      properties:
        event_description:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/str'
        event_end_time:
          type: string
          format: date-time
        event_id:
          $ref: '#/components/schemas/str'
        event_location_lat:
          type:
          - number
          - 'null'
          format: double
        event_location_lon:
          type:
          - number
          - 'null'
          format: double
        event_location_name:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/str'
        event_name:
          $ref: '#/components/schemas/str'
        event_start_time:
          type: string
          format: date-time
        event_url:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/str'
        gcal_id:
          $ref: '#/components/schemas/str'
        last_modified:
          type: string
          format: date-time
    CalendarCacheUpdateRequest:
      type: object
      required:
      - updates
      properties:
        updates:
          type: array
          items:
            $ref: '#/components/schemas/_CalendarCacheRequest'
    CalendarUpdateRequest:
      type: object
      required:
      - updates
      properties:
        updates:
          type: array
          items:
            $ref: '#/components/schemas/_CalendarListWrapper'
...
    GcalEventID:
      type: object
      required:
      - gcal_id
      - event_id
      properties:
        event_id:
          $ref: '#/components/schemas/str'
        gcal_id:
          $ref: '#/components/schemas/str'
...
    Pagination:
      type: object
      required:
      - limit
      - offset
      - total
      properties:
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...
    str:
      type: string